cr-restore: skip current from search in collect_child_pids
In this loop we are collecting children of tasks which will die and
their children will reparent to init (us). There is no point to check if
we ourselfs will die and our children will reparent to us, it does not
work like this. Though our children will be still collected but for
other reason, it is done in the end of function separately.

https://jira.sw.ru/browse/PSBM-107418

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>