pstree: prepare a link for each item to it's child subreaper
It is an optimization so that it would be easier to find subreapers for
tasks, will be used in next patch.

Logic is quiet simple: if we are pidns init we ignore any subreapers in
ancestors, else we look on our parent: if parent is subreaper it is our
subreaper (if our children will die they will reparent to it), else
if parent has some other task as it's subreaper this task is also a
subreaper for us.

https://jira.sw.ru/browse/PSBM-107418

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>