freezer: wait freezer.state changes to FROZEN after writing
The original problem is that there can be a tiny race window in
pstree_switch_state() between freezing the freeze_cgroup back in
freezer_restore_state() and detaching from processes in freezer_detach().

And if vzctl snapshot calls criu dump on frozen cgroup it expects that
processes does not unfreeze during the process even for a small amount of time
because if they do the ploop snapshot which will be collected after criu dump
can change and be inconsistent with dump. (note: then processes are stopped by
ptrace they are can be thought of as frozen in this aspect)

So to remove this possible race we should wait for cgroup to finish freezing.

Cut the common part of waiting for cgroup to become FROZEN from
freeze_processes to __wait_freezer_state_change.

Fix multiple nits in freeze_processes: fix nr_attempts calculation, remove the
use of "magic" numbers from time conversion, remove excess state
initialization, properly handle unfreeze on error path (should not be
recursive), try unfreeze on get_freezer_state error path and make the actual
number of attempts be equal to nr_attempts instead of nr_attempts+1.

Note __wait_freezer_state_change is done the way that it's number of iteration
is integrated to the outer number of iterations in freeze_processes.

https://jira.sw.ru/browse/PSBM-107938

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

rebase to 3.15 mainstream:
reuse freezer state helper functions from mainstream

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>