vz8: use legacy iptables everywhere
Binaries of iptables now all work through nft and thus can't setup iptables
rules. We have iptables network lock/unlock and we need to c/r containers
iptables rules (because it can be Centos7 without default nft).

So let's do:

- use legacy iptables-restore in network_(un)lock_internal and scripts

This is not strictly required because ipt-style rules are implicitly converted
to nft-style and work, but to be explicit lets use legacy for ipt-style.

- use legacy iptables-save/restore in iptables dump/restore

- use legacy iptables in kerndat xtlock check and modules preload

https://jira.sw.ru/browse/PSBM-120728

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>