sk-netlink: fix order between restoring queue and file params
We have path which makes netlink sockets non-blocking to be able to
restore their queue:

  +-> open_netlink_sk
    +-> restore_netlink_queue
      +-> restore_sk_queue
        +-> restore_prepare_socket enable O_NONBLOCK

But we restore file params (including O_NONBLOCK flag) just before that,
this looks like a wrong order. So lets fix it.

Also we don't really dump status flags for netlink sockets, so lets also
do it. In case of old image now we would restore all sockets with zero
flags instead of all non-blocking as it was before this patch.

This problem leads to kubernetes weave-net network do die on EAGAIN
returned from socket which is expected to be blocking. And also weave
can't restart, as it can't remove weave and another veth devices as all
it's netlink controlling sockets are broken.

Fixes: 870a647d5f9c ("vz7: netlink: restore data in a received queue for
netlink sockets")
Note: 6ca2cb1a0e88 ("vz7: netlink: restore queues and only then restore
opts")

https://jira.sw.ru/browse/PSBM-120924
Reviewed-by: Andrey Zhadchenko <andrey.zhadchenko@virtuozzo.com>
Reviewed-by: Alexander Mikhalitsyn (Virtuozzo) <alexander@mihalicyn.com>
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

rebase to mainstream 3.18:
flags hunk is not needed anymore as it was fixed in mainstream

https://jira.sw.ru/browse/PSBM-144416
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Netlink packet queue