binfmt_misc: isolate binfmt_misc dump logic from mountinfo/mountpoint dump.
binfmt_misc filesystem is a special case in criu, primarily because it
does not loose state, when unmounted. This means it does not fit into
criu's mountpoint dump pipeline, because criu only dumps those
filesystems that it knows about from parsing /proc/pid/mountinfo.
At dump if unmounted binfmt_misc will not show up in mountinfo and thus
will not be added to a mountree and mountlist that is formed at
collect_mountinfo function.
That is why criu has a special function that mounts binfmt_misc if it
was not present in mountinfo and then dumps it's contents.
The problem is that traditional target location is inside of a dumped
container and this target location is prone to "mount trap". Also we
don't want to do technical mounts inside of a container without really
special need.
In this patch binfmt_misc mount is done outside of a container. Also
binfmt_misc dump is removed from generic pipeline and is only dumped
in one single code location.

Feature: Binfmtmisc

https://jira.sw.ru/browse/PSBM-96362
Reviewed-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>
Signed-off-by: Valeriy Vdovin <valeriy.vdovin@virtuozzo.com>