restore: disable clone3 + set_tid code
The clone3 + set_tid case does not fit into nested pidns and userns case
because: 1) clone3 requires to be ns_capable(CAP_SYS_ADMIN) in each owner
userns of each pidns of pid array; 2) to create init of the pidns via
clone(CLONE_NEWPID) one should be in owner userns of the pidns (almost
the same for unshare(CLONE_NEWPID) + clone()). So (1) contradicts with
(2) and we can't have both of them, but we need, and we can do it with
old known ns_last_pid.

Integrating with clone3 + set_tid for a simple case in scope of rebase
looks like a worthless path, so lets just disable it.

https://jira.sw.ru/browse/PSBM-121533

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>