memfd: use PROC_SELF instead of getpid in __open_proc
This looks better for me, should be no functional change.

We need to populate this new SELF service fd in populate_pid_proc, so
that it is later available in protected context. Also leave PROC_SELF
service fd open in prepare_fds because it is safe to have service fds
open on files restore stage.

Another implication of this is nested pid namespaces, when we will
support them "__open_proc(getpid()...)" will try to open file of the
process which has the same pid but in NS_ROOT pidns, which is bad.

See also aa2d92082 ("files: use PROC_SELF when a process accesses its
/proc/PID")

https://jira.sw.ru/browse/PSBM-121533

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>