util: move open_proc_self_fd to service_fd
We need this to avoid conflicts with file descriptors which has to be
restored. Currently open_proc PROC_SELF is not used during restoring
file descriptors, but we are going to use it for memfd restore.

Note: in get_proc_fd we don't close service fd if we detect service fd
of other task it will be replaced in open_pid_proc anyway, and e.g. in
protected sfd context we can't close or open sfd but can replace it
without any problems.

While on it also add FIXME because the check in get_proc_fd is error
prone in some rare cases.

Also don't close /proc/self service fd in prep_unix_sk_cwd as it can be
called from protected context and there is no much point of closing it
anyway.

https://jira.sw.ru/browse/PSBM-121533

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>