servicefd: fully cleanup parent's shared fd table on service fd update
Before this patch while moving parent's service fds to a new place we
didn't close what previously was in this old place, this way if we try
to open some service fd which has this undesired open fd lefover in
it's place we will get "busy target" error from install_service_fd.

It's a preparation to make /proc/self to be a service fd, if we do so
without this patch the problem with busy target is triggered on zdtm.

https://jira.sw.ru/browse/PSBM-121533

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>