restore/ve: do ve start from container init task
In vz8 kernel we don't have a way to set ve.state="START pid" so only the init
of container can start ve now. So we can't start ve from scripts now, only from
criu code in vz8.

But this requires a corrsponding change in libvzctl (delete starting ve
from libvzctl scripts), to be able to use the same libvzcl in both vz7
and vz8 (as we currently do) we need to port this change to vz7 criu
too.

So add start_ve() helper and call it just after "setup-namespaces" part of
libvzctl vz-rst-action script was running but from init task not criu task.

https://jira.sw.ru/browse/PSBM-123162
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: VE