build: add ability to override CRIU_VERSION from rpmbuild specs
Now we see only mainstream criu version in criu logs:

  []# criu --version
  Version: 3.12

If we add an ability to override it we would see more informative:

  []# criu/criu --version
  Version: 3.15.1.4

If we build criu with "CRIU_VERSION=3.15.1.4". It can be very useful as
we would not need to use rpm to get version.

See also spec patches for vz8/vz7:
https://git.sw.ru/projects/VZSPECS/repos/criu/pull-requests/10/overview
https://git.sw.ru/projects/VZSPECS/repos/criu/pull-requests/11/overview

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Miscellaneous