files-reg: show sdev of the file we failed to lookup mount for
When we see "Can't lookup mount" error in report there is not always
enough info about why mount was not found. There can be at least 4
cases:

- file from detached mount
- file from host's mount
- special fd from pseudo (invisible) mount
- overmounted "/" makes some mounts in mountinfo invisible for criu

So let's print info about the device the file is from. It can help
investigation.

Also show some available bindmount (if have any) of this sdev as a hint,
because device number is in "user" format and devices of mounts are in
"kernel" format in logs, and converting them is a bit of a pain.

https://jira.sw.ru/browse/PSBM-124422
https://jira.sw.ru/browse/PSBM-124576

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>