zdtm: add file_fown_reuse to check dead owner pid reused
There are three cases if file has file owner:

1) Owner is still alive - so we need to restore it after c/r
2) Owner is dead already - we can't restore it after c/r, pid just would
not be found.
3) Owner is dead, but some other task reused it's pid - we should not
restore this poor task as a file owner because it can later get
unexpected signals.

So let's add a testcase for case (3). It would not pass on kernels
without a speciall fix, because there is no way to determine if it's (1)
or (3) without the fix.

https://jira.sw.ru/browse/PSBM-124623

Reviewed-by: Alexander Mikhalitsyn <alexander.mikhalitsyn@virtuozzo.com>
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>