criu: rework is_zdtm_run
Currently vz criu is used to dump either ve or zdtm tests. Unfortunately we
use is_zdtm_run as indicator that we shouldn't dump ve-related things. In
general this is not correct. Lets use recently added opts.ve to decide if we
dump such things. With this ZDTM_RUN environmental variable and
is_zdtm_run() func is no longer needed.

https://jira.sw.ru/browse/PSBM-122120
Reviewed-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>
Signed-off-by: Andrey Zhadchenko <andrey.zhadchenko@virtuozzo.com>