mount: don't mount from external source if already have a bind
If we have several external mounts of same superblock and one can be
bind from another, there is no need to handle both of them as external
bindmounts. We can just handle second mount as normal "internal" bind.

More over this will fix one rare bug that external-dev mount was
detected as a bind of some other mount and thus got to do_bind_mount_v2,
but it was handled as non-dev external bindmount which is wrong (as
external-dev can be only created from do_new_mount_v2).

Reviewed-by: Alexander Mikhalitsyn <alexander.mikhalitsyn@virtuozzo.com>
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>