cgroup/device: update device major/minor in device.list
If one of devices in device.list content from images is an external
device there is a possibility that it would change device id between
supend and resume. For instance our ploop virtual device generates new
ids on creation.

So we must update ids in device.list file content.

Note: we only update single device rules as device group rules with "*"
are most likely unchanged.

Note: in future we might want more effective algorithm to deal with
strings here, copying lines mostly without replacement from value to
result is not so fast.

Note: all options in images are restricted by size 1024 characters, so
it should be enough to have 2*1024 for updated options.

https://jira.sw.ru/browse/PSBM-123668

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>