devices: resolve new device numbers for each device
When we dump a device it is provided via:
  --external "dev[major/minor]:key"
option on dump, and on restore it should have corresponding
  --external "dev[key]:/dev/path"
option to resolve old device to a corresponding new device by key.

Previousely we've only used those mappings for restoring device external
mounts, but we can reuse them to update major/minor pair to new values
of the replacing device. Let's get new major/minor pair from device path
provided in options and put it on device structure.

v2: remove excess +1 in len

https://jira.sw.ru/browse/PSBM-123668

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>