zdtm: make cgroup04 also check cpuset special props
Previousely in these patchset we've changed --manage-cgroups=full to
actually restore special props, which was not true previousely.

These cgroup04 test which uses --manage-cgroups=full needs correction,
as now all special props in all dumpee tree cgroups are c/r-ed and the
test is in root cpuset cgroup that's why we try to c/r root
cpuset.cpus and cpuset.mems which is not permitted. Let's change the
test, so that it enters non-root cpuset cgroup directory and also we can
change that cpuset.cpus and cpuset.mems are actually c/r-ed this way.

Also improve cleanups, so that all cgroup controllers are handled.

So now we can reenable cgroup04 test.

https://jira.sw.ru/browse/PSBM-123668
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: ZDTM