nft: run nft binary from rpm414x64 chroot
The original problem is that docker in centos8 nft-based container can
create via ipt->nft conversion so called "fib" rules, our nft binary on
vz7 host (nftables-0.8-14.vl7) is not capable to properly dump those
rules, that's why we need to use a "newer" binary, similar as we have in
centos8 template (nftables-0.9.3-12.el8). We can't simply update nft on
host because it will also affect everything else including iptables.

So we've decided to have a special chroot environment rpm414x64 in
/vz/pkgenv/ so that criu can use nft binary from it with all needed
dependencies.

Also for nft ruleset restore we need to have /proc mounted in this
chroot so that we can access /proc/self/fd/0. That's why simple chroot
does not work for us. We create mntns with private propagation, prepare
our future chroot directory, mount proc inside and do pivot_root to it.
Also we call everything under call_in_child_process that's why fd tables
are shared so we can use prepared fds and when our task will exit this
temporary mntns and it's temporary mounts would be cleaned up.

Note1: On vz8 we have same /vz/pkgenv/rpm414x64.

Note2: Proc is bindmounted recursively as on restore /proc has MNT_LOCKED
children mounts which can't be umounted from it due to change of userns.

Note3: We can actually reuse the chrooted mntns for dumping nft rules in
each netns instead of recreating it each time. We may optimize it in
future.

https://jira.sw.ru/browse/PSBM-125002
Reviewed-by: Alexander Mikhalitsyn <alexander.mikhalitsyn@virtuozzo.com>
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Nftables