cgroup: skip cgroup namespace restore when cgset is not dumped
If on dump all dumped tasks are in same cgset with criu we don't dump
cgsets and thus cgroup namespaces and rely that on restore criu would
prepare proper cgset/cgns for us.

This is needed to fix zdtm tests as they mostly run in same cgroups with
criu without any cgroup namespace. We saw errors like this:

  Error (criu/cgroup.c:1239): cg: No set 1 found

Note: This skip is the same to what we do in restore_task_cgroup.

https://jira.sw.ru/browse/PSBM-123431

Fixes: 4ab1a32cf ("cgroup/restore: split prepare_task_cgroup code into
two separate functions.")
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>