cgroup: added command line flag to force creation of cgroup namespace
Not long ago new cgroup namespace creation was optional and it was deduced
from the dump weather or not we should recreate previously dumped cgroup
namespace. This was done by examining cgns_prefix value above.
Later, VZ8 kernel was modified to REQUIRE that init task placed in a
separate new cgroup namespace before container start.
The changed behaviour of VZ8 kernel is a problem, because if no
cgns_prefixes are present in the dump, then unshare below will be skipped and
VZ8 will refuse to start a container.
Simple solution is to let unshare be called always for all kernels that
support cgroup namespace. But this solution brings regression both known
and undiscovered to other possible cases in VZ7 and zdtm.

A nicer more stable solution is a command line option that forces new cgroup
namespace creation, this way all logic remains and vzctl can deduce from it's
context wheather to set this option or not.

Reviewed-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>
Signed-off-by: Valeriy Vdovin <valeriy.vdovin@virtuozzo.com>

Feature: Cgroup namespace force creation for ve