jenkins: run ve flavor tests without extra pidns in a separate job
The script test/jenkins/run_ct creates new pidns for running zdtm. For ve
flavor we can't afford it as starting ve is not supported on too deep nested
pidns (depth >2).

So let's skip ve flavour in normal run (criu.sh) and prepare to run it in a
separate Jenkins job.

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: VE