cgroup: at restore use fallback values for cpuset.cpus/mems
If container is being migrated from one machine to another with a
smaller amount of CPU cores, criu will fail to restore cgroup property
'cpuset.cpus'. It will then fail the whole migration.
Ex src node has 8 cores, and cpuset.cpus will have value '0-7' by default. This
value is too big for machine with say 4 cores, it's maximum range would be
'0-3'. Writing '0-7' will result in -ERANGE. Same thing is for 'cpuset.mems'.

It is solved in this commit by introducing a fallback logic for writes
specificly for this 2 parameters "cpuset.cpus" and "cpuset.mems". To implement
this fallback logic it's necessary to first read out the 'allowed' values for
all the properties of our interest, so that later we could write them instead
of the failed ones. We read this values from the root-level cgroups.

https://jira.sw.ru/browse/PSBM-127807

Reviewed-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>
Signed-off-by: Valeriy Vdovin <valeriy.vdovin@virtuozzo.com>

rebase to mainstream 3.18:
merge debug message fixup
https://src.openvz.org/projects/OVZ/repos/criu/commits/797aa7ea0

https://jira.sw.ru/browse/PSBM-144416
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Cgroup