ps-socket: sanity check that given fd is actually valid
We see that if fd was not initially open it can lead to situations where
this fd is reused with some criu open and later page server may fail on
this fd of other type (non-socket).

v2: instead of checking via /proc/self/fd/ just do a fstat on fd, plus
we can check if it's a socket or not this way.

https://jira.sw.ru/browse/PSBM-102629
https://jira.sw.ru/browse/PSBM-130507
Reviewed-by: Alexander Mikhalitsyn <alexander.mikhalitsyn@virtuozzo.com>
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Miscellaneous