cgroup: skip beancounter cgroup restore if it's not supported
On vz7 we have beancounter cgroup and we dump it into images, on vz8 we
don't have it anymore so we can't restore it. But it should be safe to
just skip it's restore because we actually does not dump much in
beancounter cg anyway, only general stuff: cgroup.clone_children,
notify_on_release, cgroup.procs and tasks.

https://jira.sw.ru/browse/PSBM-131412
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Beancounters