config: add --ve-clock-fallback option
This option should be used on restore, if we restore on vz8/vz9 with
time-namespaces enabled in kernel, we need to restore container in a
time-namespace. If dump was from vz7 with no time-namespace data in it
we should fallback using our old ve.clock_* values, saved by libvzctl
for us.

Usage example:
--ve-clock-fallback monotonic:15.687969006,boottime:15.687557105

Feature: Time namespace to ve clock fallback

https://jira.sw.ru/browse/PSBM-134480
Reviewed-by: Alexander Mikhalitsyn <alexander.mikhalitsyn@virtuozzo.com>
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>