infect: add SIGTSTP support
Add SIGTSTP stop signal dumping and restoration. Add a field
for this in the image, save it only if a task is in the stopped
state.

Restore task state by sending desired stop signal if it is present
in the image. If it isn't then send the default SIGSTOP.

Update compel samples with SIGSTOP and SIGTSTP signals support.

https://jira.sw.ru/browse/PSBM-127306
Signed-off-by: Yuriy Vasiliev <yuriy.vasiliev@virtuozzo.com>