sk-unix: improve detection of mnt_id of bindmounted sockets for old images
If we create bindmounted socket on bindmount, this bindmount will be
counted as one of unix bindmounts because in sk_has_bindmount() after
checking all of vz_bind_mnt_ids[] we check ue->mnt_id itself as in old
images this field contained bindndmount id.
The correct way is to check this only if n_vz_bind_mnt_ids == 0. Otherwise
we will try to make unix socket on a mount which is not yet mounted

https://jira.sw.ru/browse/PSBM-135398
Reviewed-by: Alexander Mikhalitsyn <alexander.mikhalitsyn@virtuozzo.com>
Reviewed-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>
Signed-off-by: Andrey Zhadchenko <andrey.zhadchenko@virtuozzo.com>

Feature: Unix sockets