filesystems: allow dumping container with nfsd mounts
Criu can't dump nfs server, for instance:

- nfsd/rpcpipefs contents is not restored
- opened /proc/<pid>/net/rpc/auth.unix.ip/channel is not restored

This is a workaround to support nfs server migration. Idea is to stop
nfs server before suspending container and restart it after resuming it.
As nfs server is stateless, after restart all nfs clients of this server
would reconect flawlessly, so it should not be much difference between
this w/a and honest migration.

So we would stop nfs server before dump, so nfsd mount should not be
used at the time of dump, so we only need to place nfsd mount in the
topologically the same place after dump and don't even handle open
files.

https://jira.sw.ru/browse/PSBM-138439
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: NFSD