action-scripts: call dump-alive when dumped task is left running
This hook can be useful to revert any changes to the dumped container
done in pre-dump script. For instance if the container can't be dumped
due to some known service using some known undumpable resource, we can
just stop this service in pre-dump script and then restart it from
post-resume script (a workarond for stateless services). But in this
case if actual errors in criu dump arise we should be able to revert the
service stop and let the container continue running, we would do it from
this dump-alive action script hook.

https://jira.sw.ru/browse/PSBM-138439
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: NFSD