irmap: don't fail early if failed to resolve path on pre-dump
Irmap is a cache to make resolving inode+sdev pairs to paths easier
(e.g. for inotify watches), irmap cache is collected on pre-dump when
tasks are not frozen as it is a time consuming operation and we want to
get lower downtime of migration.

In this patch we want to optimize one special case where criu can not
resolve the path for an inotify watch on rpc_pipefs open by nfs server
related service nfsdcld and it fails pre-dump.

We can safely skip failing here, just don't add an irmap if it was not
resolved, this effectively delays the fail to a later stage, if on dump
we still face this unresolvable inotify watch in:

  dump_one_inotify
    check_one_wd
      check_open_handle

We would fail.

Though if nfsdcld had stopped for some reason (e.g. our nfs server
migration w/a temporary stopped it, or container user luckily had
stopped it) and we don't have this unresolvable inotify watch anymore,
we are free to dump container without any errors.

https://jira.sw.ru/browse/PSBM-140177

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>