net: dump nft rules via both nft and iptables-nft
We see nft rules with matches like "physdev" or "match-set" which can be
created by iptables-nft but can not be dumped/restored by nft utility.
We also see custom nft tables (not "filter", "nat" or "mangle", etc.)
created by nft are not dumped by iptables-nft.

The idea of this patch is to dump nftables rules both with nft and
iptables-nft, and on restore try to restore them both.

The --nftables-mode=ipt/nft/all option is added to switch to nft only or
itpables-nft only mode (default is all).

https://jira.sw.ru/browse/PSBM-138902
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Nftables