sk-unix: allocate FdinfoEntry with shmalloc in get_fle_for_task
We already do so in other branch of if in dup_fdinfo, let's then do it
symmetrically. We later put pointer to this FdinfoEntry in collect_fd_to
int alloc_fle to fdinfo_list_entry->fe so it can be later used from
other places and if one thread modifies it we would not see modification
in other threads.

Do we really need to see these modifications? - It's Mistery, but better
be on the safe side.

https://jira.sw.ru/browse/PSBM-143751

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>