sk-unix: resolve fake queuers vs scm fds in unix sockets fallacy
Long story short: we face standard "which was first chicken or egg?"
problem here.

1) In add_fake_unix_queuers() for each unpared established unix socket
we want to add a "fake" queuer. And we want to decide at this point
which task should create this queuer and we just try to copy the task
from original socket. But original socket can have no task yet too,
because it can have no open fds, instead it is residing inside other
unix socket.

2) In prepare_scms we look through the packets from all unix sockets and
detect file descriptors in them.

3) We need to add those packets in special queuer's list so that queuer
can understand when all files we need to put in it are already created.
So it means that we need to have "fake" queuers present at this point.

4) At the same time it is also a point to understand which task should
create those descriptors from unix sockets and we just try to copy the
task from the queuer who would put them to unix socket.

You can imagine that queuers should be created (1) before we try to copy
task from them (3) and their peer socket may only give the task to
queuers (1) after it copyes the task from it's ancestor unix socket (4).

So let's fix it with:

- Split "fake" queuer owner task detection into separate step after
handling scm

- When handling scm, if queuer has no owner task, take it from original
socket

https://jira.sw.ru/browse/PSBM-143751
Fixes: 5cd7092fd ("sk-unix: make add_fake_unix_queuers earier and rework find_queuer_for")
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Unix sockets