zdtm: Fix setting up zdtm test env variables
This condition evaluates to false not only in case when env variable is
None, but also when env is an empty dict. So, in some cases env could be
an empty dict (for example, when we use flavor: 'h', flags: 'suid' and
freezecg: 'zdtm/:t').

Empty env leads to errors like:

===================== Run zdtm/static/cgroup_freezer in h ======================
Start test
Test is SUID
make: sh: No such file or directory
make: uname: No such file or directory
make: uname: No such file or directory
make: make: No such file or directory
make: *** [Makefile:731: ../lib/libzdtmtst.a] Error 127

To fix this, let's check if env is an instance of dict class.

https://jira.sw.ru/browse/PSBM-144187

Reviewed-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>
Signed-off-by: Yuriy Vasiliev <yuriy.vasiliev@virtuozzo.com>