util: add __get_relative_path version with const arguments and return
We want to use get_relative_path in mount_resolve_path, so that we need
to either remove const or propagate const to get_relative_path. Both
variants require a lot of collateral changes, so let's just make const
version and wraper to it which casts away the constness.

rebase to mainstream 3.18:
rework: instead of removing const add it to get_relative_path
https://src.openvz.org/projects/OVZ/repos/criu/commits/9d6a63ead

https://jira.sw.ru/browse/PSBM-106869
https://jira.sw.ru/browse/PSBM-144416
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Mounts