sk-queue: order unixsk sent to other unixsk after it
Here is how we dump nested unix sockets inside unix sockets:

  +-> dump_one_unix_fd         <-+
    +-> dump_sk_queue            |
      +-> dump_sk_queue_packet   |
        +-> dump_packet_cmsg     |
          +-> dump_one_file      |
            +-> do_dump_gen_file +
        +-> queue_packet_entry
          +-> list_add_tail(packets_list)

So in the images more deeper unix socket goes strictly before the more
shallow one.

Before this patch on restore we also had the same reverse order due to
list_add_tail used when adding to packets_list. Because of it we first
handle deeper sockets in prepare_scms and later shallow ones.

This is conterintuitive because in case of nested unix sockets without
any actual open fds on them we need to find out which task should create
them (file_master()->task) and in unix_note_scm_rights we determine this
owner from queuer which can also have it yet unset because it would be
handled later in prepare_scms.

Let's reverse the walk order on restore in prepare_scms.

Note: we can't just reverse the order in list as we rely in
restore_sk_queue that packets of one socket are in original order.

https://jira.vzint.dev/browse/PSBM-143751
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Unix sockets