mount: introduce helper mount types
We plan to add more types of helper mounts. There are already a lot
of them and there is no easy way to distinguish types. So let's
explicitly mark them.

At the same time make the check in clean_cr_time_mounts cleaner, as we
are interested only in cr_time mounts there.

https://jira.vzint.dev/browse/PSBM-147120
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Mounts