mount: search covering mounts for sharing groups
In general, if we ommit corner cases like multiple external mounts with
same sharing group, to restore all mounts of one sharing group we should
have a mount ("wide" mount) in it with root covering all other mounts of
this sharing group and all mounts of it's descendants.

By searching covering mounts for each sharing group we can understand if
this sharing group has such a single "wide" mount or not.

https://jira.vzint.dev/browse/PSBM-147120
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Mounts