mount: add wide mounts for sharing groups where required
In general, if we ommit corner cases like multiple external mounts with
same sharing group, to restore all mounts of one sharing group we should
have a mount ("wide" mount) in it with root covering all other mounts of
this sharing group and all mounts of it's descendants.

When we have sharing group without "wide" mount we can restore it by
creating helper "wide" mount.

This helper "wide" mount:
- is added outside of mount namespaces of container in root yard,
- has plain path "<yard>/sg-<master_id>-<minor_id>-<bind_mnt_id>"
- made to be only bindmounted from other mount of same sb,
- has custom dir/file detection from bindmount
- has multiple fields in mountinfo left blank

Here is how we decide that we need "wide" mount:

Traverse all sharing groups recursively child before parent:

a) If a sharing group has single covering mount, we check if this mount
is from the sharing group, if yes - we don't need to do anything, if no
- we add helper mount with root equal to covering mount root to this
sharing group.

b) If a sharing group has two or more covering mounts, we lookup in
ancestor sharing groups a covering mount which covers all covering
mounts of current sharing group:

  c) if found - we add helper mount with root equal to found covering
  mount root to this sharing group and update covering mounts for this
  sharing group and all ancestors

  c) if not found - we search the same in superblock covering mounts

    d) if found sb covering mount - we add a helper mount based on it
    and update covering mounts for this and ancestor sharing groups

    e) if not found sb covering mount:

      f) if also sharing group has no parent or has shared_id - fail

      g) else sharing group is a slave of multiple external sources - so
      we add helper mounts for each of coverings mounts of this sharing
      group and prey =)

https://jira.vzint.dev/browse/PSBM-147120
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Mounts