mount: introduce covering mounts functionality
This is designed for chosing, for the set of mounts, a minimal set of
mounts which covers all other mounts in terms of mount's root. Meaning
that any particular mount from set has a mount in covering set there
covering mount root is a subpath of the particular mount root.

For instance, if we have mounts with roots:

    /a
    /a/b
    /c/d
    /c/d/e
    /f/g
    /f/h

covering mounts for them are:

    /a
    /c/d
    /f/g
    /f/h

https://jira.vzint.dev/browse/PSBM-147120
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Mounts