soccr: do not disable tcp repair when sending unsent data
Disabling tcp repair here can lead to tcp_reset() in kernel for the
socket and after that we would get ECONNRESET and EIO when sending to
this socket.

This is more a work around as I don't fully understand why we see this
problem, it seemed that nothing changed in this area in 3.18 rebase.

https://jira.vzint.dev/browse/PSBM-147346
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Tcp Repair