sk-unix: ignore closed or non-empty in-flight sockets
Here is another hacky patch, it basically reverts mainstream patch:
dca55d281 ("criu: fail migration if data was sent to an in-flight socket")
Leave the error messages about the problem to still be aware of it.

It is error-prone to do so, but we lived without this safety check for
some time, so we can fix it properly later and revert it to unblock
tests for now.

https://jira.vzint.dev/browse/PSBM-147983
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Unix sockets